formatted

+++ model/mlp_in_keras/mlp_in_keras.py
@@ -1,8 +1,7 @@
-# Ref: 
+# Ref:
 # https://machinecurve.com/index.php/2019/07/27/how-to-create-a-basic-mlp-classifier-with-the-keras-sequential-api
 
 # Imports
-import tensorflow
 from tensorflow.keras.datasets import mnist
 from tensorflow.keras.models import Sequential
 from tensorflow.keras.layers import Dense
@@ -21,8 +20,8 @@ X_train = X_train.reshape(X_train.shape[0], feature_vector_length)
 X_test = X_test.reshape(X_test.shape[0], feature_vector_length)
 
 # Convert into greyscale
-X_train = X_train.astype('float32')
-X_test = X_test.astype('float32')
+X_train = X_train.astype("float32")
+X_test = X_test.astype("float32")
 X_train /= 255
 X_test /= 255
 
@@ -32,18 +31,18 @@ Y_test = to_categorical(Y_test, num_classes)
 
 # Set the input shape
 input_shape = (feature_vector_length,)
-print(f'Feature shape: {input_shape}')
+print(f"Feature shape: {input_shape}")
 
 # Create the model
 model = Sequential()
-model.add(Dense(350, input_shape=input_shape, activation='relu'))
-model.add(Dense(50, activation='relu'))
-model.add(Dense(num_classes, activation='softmax'))
+model.add(Dense(350, input_shape=input_shape, activation="relu"))
+model.add(Dense(50, activation="relu"))
+model.add(Dense(num_classes, activation="softmax"))
 
 # Configure the model and start training
-model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
+model.compile(loss="categorical_crossentropy", optimizer="adam", metrics=["accuracy"])
 model.fit(X_train, Y_train, epochs=10, batch_size=250, verbose=1, validation_split=0.2)
 
 # Test the model after training
 test_results = model.evaluate(X_test, Y_test, verbose=1)
-print(f'Test results - Loss: {test_results[0]} - Accuracy: {test_results[1]}%')
+print(f"Test results - Loss: {test_results[0]} - Accuracy: {test_results[1]}%")
